pyproject.toml
[project]
name = "iamcompact-nomenclature"
version = "0.9.2"
description = "Data structure definition for validation of model outputs in the HORIZON EUROPE project IAM COMPACT, using nomenclature-iamc"
authors = [
    { name = "Jan Ivar Korsbakken", email = "janko@cicero.oslo.no" }
]
dependencies = [
    "gitpython>=3.1.44",
    "nomenclature-iamc",
    "pyam-iamc>=3.0",
]
readme = "README.md"
requires-python = ">= 3.13"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipython>=8.20.0",
    "pytest>=8.0.0",
    "ipykernel>=6.29.0",
    "dvc[ssh]>=3.48.4"
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["iamcompact_nomenclature"]
